+++ crates/plugin_runtime/src/render_resource_pot/source_replacer.rs
@@ -9,11 +9,11 @@
 
 use farmfe_core::{
   config::{Mode, TargetEnv, FARM_DYNAMIC_REQUIRE, FARM_REQUIRE},
-  module::{module_graph::ModuleGraph, ModuleId, ModuleType},
+  module::{module_graph::ModuleGraph, ModuleId},
   plugin::ResolveKind,
   swc_common::{Mark, DUMMY_SP},
   swc_ecma_ast::{
-    Bool, CallExpr, Callee, Expr, ExprOrSpread, Ident, Lit, MemberExpr, MemberProp, Str,
+    Bool, CallExpr, Callee, Expr, ExprOrSpread, Ident, Lit, MemberExpr, MemberProp, Number, Str,
   },
 };
 use farmfe_toolkit::{
@@ -38,15 +38,26 @@ pub struct SourceReplacer<'a> {
   target_env: TargetEnv,
 }
 
+pub struct SourceReplacerOptions<'a> {
+  pub unresolved_mark: Mark,
+  pub top_level_mark: Mark,
+  pub module_graph: &'a ModuleGraph,
+  pub module_id: ModuleId,
+  pub mode: Mode,
+  pub target_env: TargetEnv,
+}
+
 impl<'a> SourceReplacer<'a> {
-  pub fn new(
-    unresolved_mark: Mark,
-    top_level_mark: Mark,
-    module_graph: &'a ModuleGraph,
-    module_id: ModuleId,
-    mode: Mode,
-    target_env: TargetEnv,
-  ) -> Self {
+  pub fn new(options: SourceReplacerOptions<'a>) -> Self {
+    let SourceReplacerOptions {
+      unresolved_mark,
+      top_level_mark,
+      module_graph,
+      module_id,
+      mode,
+      target_env,
+    } = options;
+
     Self {
       unresolved_mark,
       top_level_mark,
@@ -104,7 +115,7 @@ impl SourceReplacer<'_> {
   }
 
   fn replace_source_with_id(&mut self, call_expr: &mut CallExpr) -> SourceReplaceResult {
-    // require('./xxx') or require('./xxx', true)
+    // not require('./xxx') or require('./xxx', true)
     if call_expr.args.len() < 1 && call_expr.args.len() > 2 {
       call_expr.visit_mut_children_with(self);
       return SourceReplaceResult::NotReplaced;
@@ -152,7 +163,7 @@ impl SourceReplacer<'_> {
           return SourceReplaceResult::NotReplaced;
         }
 
-        if dep_module.module_type.is_script() || dep_module.module_type == ModuleType::Runtime {
+        if dep_module.module_type.is_script() {
           // println!("replace {:?} to {:?}", value, id.id(self.mode.clone()));
           str.value = id.id(self.mode.clone()).into();
           str.span = DUMMY_SP;