+++ packages/core/src/utils/index.ts
@@ -9,3 +9,4 @@ export * from './publicDir.js';
 export * from './rebase-url.js';
 export * from './plugin-utils.js';
 export * from './cacheDir.js';
+export * from './dynamic-resources.js';